feat(renderer): Add features and improve logic
Expand "switchOnOff" logic to make various checks and stay disabled for 2 seconds everytime it is clicked to prevent spamming the button ( could generate unwanted behaviour like switch OFF with light ON, ecc. ).
The function will still contain checks even with the timer to still solve possible issues ( if light is ON and switch is OFF then turn switch ON, ecc. ).
The light's state will be changed only after TPLightStatus() has finished to prevent changing state while the light is still managing the previous command ( changed state would not be registered by the device ).

Set brightness value in lightStatus() function to be "off" if the brightness value is undefined (this happens when we try to retrieve the brightness value with the light turned off, it just gives the on_off value) or "brightness" if the value exists.

Add an EventListener("input") for when the range slider is interacted with. This will keep updating the light while the user moves the pointer. As above, light's value is updated only after Status() is done.

Add an EventListener("change") to update the value once the mouse click is released (May be redundant)

Commented unused console logs